requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pytest
httpx
pytest-asyncio
aiohttp